+++ lib/src/easy_resend_base.dart
@@ -5,13 +5,16 @@ import 'package:easy_resend/src/models/attachment.dart';
 import 'package:easy_resend/src/models/errorResponse.dart';
 import 'package:http/http.dart' as http;
 
+/// A class for sending emails by Resend.
 class EasyResend {
-  // Private constructor
+  /// Private constructor for creating an instance of EasyResend.
   EasyResend._(this._apiKey);
-
   final String _apiKey;
   static EasyResend? _instance;
 
+  /// Initializes the EasyResend singleton instance with the provided API key.
+  ///
+  /// Throws an exception if the instance has already been initialized.
   static void initialize(apiKey) {
     if (_instance != null) {
       throw Exception('The instance can only be initialized once');
@@ -20,7 +23,9 @@ class EasyResend {
     }
   }
 
-  // Singleton instance retrieval method
+  /// Retrieves the singleton instance of EasyResend.
+  ///
+  /// Throws an exception if the instance has not been initialized.
   static EasyResend getInstance() {
     if (_instance != null) {
       return _instance!;
@@ -29,6 +34,23 @@ class EasyResend {
     throw Exception('Instance not initialized');
   }
 
+  /// Sends an email using the Easy Resend API.
+  ///
+  /// Required parameters:
+  /// - [from]: The email address of the sender.
+  /// - [to]: List of email addresses of the recipients.
+  /// - [subject]: Subject of the email.
+  ///
+  /// Optional parameters:
+  /// - [text]: Plain text content of the email.
+  /// - [attachments]: List of attachments to include in the email.
+  /// - [bcc]: List of email addresses to send a blind carbon copy (BCC).
+  /// - [cc]: List of email addresses to send a carbon copy (CC).
+  /// - [html]: HTML content of the email (if provided, it will be used instead of [text]).
+  ///
+  /// Returns the ID of the sent email if successful.
+  ///
+  /// Throws an exception if an error occurs during the email sending process.
   Future<String> sendEmail({
     required String from,
     required List<String> to,
@@ -59,8 +81,10 @@ class EasyResend {
       body: jsonEncode(body),
     );
     if (response.statusCode == 200) {
-      Map<String, String> body = response.body as Map<String, String>;
-      return body['id']!;
+      String body = response.body;
+
+      Map<String, dynamic> bodyMap = jsonDecode(body);
+      return bodyMap['id']!;
     } else {
       final error = ErrorResponse.fromMap(jsonDecode(response.body));
       throw error.returnMessage();