+++ test/easy_resend_test.dart
@@ -1,25 +1,23 @@
-import 'dart:developer';
-
 import 'package:easy_resend/src/easy_resend_base.dart';
 import 'package:test/test.dart';
 
 void main() {
   group('A group of test', () {
     setUp(() {
-      EasyResend.initialize('re_9CT38ccy_NFfrVemifkCHUNLpMFWjqzHM');
+      EasyResend.initialize('your_api_key_here');
     });
 
     test('Without Attachment', () async {
       EasyResend easyResend = EasyResend.getInstance();
 
       String id = await easyResend.sendEmail(
-        from: 'Info <info@enviarmasivo.com>',
-        to: ['romand.buitrago@gmail.com'],
+        from: 'onboarding@resend.dev',
+        to: ['username@mail.com'],
         subject: 'Test Package',
         text: 'There you go!',
       );
 
-      log(id);
+      print(id);
     });
   });
 }